+++ src/or_error.mli
@@ -56,8 +56,10 @@ val ok_exn : 'a t -> 'a
 (** [of_exn ?backtrace exn] is [Error (Error.of_exn ?backtrace exn)]. *)
 val of_exn : ?backtrace:[ `Get | `This of string ] -> exn -> _ t
 
-(** [of_exn_result (Ok a) = Ok a], [of_exn_result (Error exn) = of_exn exn] *)
-val of_exn_result : ('a, exn) Result.t -> 'a t
+(** [of_exn_result ?backtrace (Ok a) = Ok a]
+
+    [of_exn_result ?backtrace (Error exn) = of_exn ?backtrace exn] *)
+val of_exn_result : ?backtrace:[ `Get | `This of string ] -> ('a, exn) Result.t -> 'a t
 
 (** [error] is a wrapper around [Error.create]: