+++ CHANGES.md
@@ -37,6 +37,8 @@
 - [Int.Hex.of_string] and friends no longer silently ignore a suffix
   of non-hexadecimal garbage.
 
+- Added `?backtrace` argument to `Or_error.of_exn_result`.
+
 ## v0.11
 
 - Deprecated `Not_found`, people who need it can use `Caml.Not_found`, but its